% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{callAPI}
\alias{callAPI}
\title{Handle interactions with the Uber API}
\usage{
callAPI(cmd, version, params = NULL, method = "GET")
}
\arguments{
\item{cmd}{Specifies the endpoint.}

\item{version}{The version of the API for particular endpoint.}

\item{params}{Data passed with request.}

\item{method}{Type of request.}
}
\description{
Handle interactions with the Uber API
}
\keyword{internal}

