% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_estimate_price}
\alias{uber_estimate_price}
\title{Price estimate}
\usage{
uber_estimate_price(start_latitude, start_longitude, end_latitude,
  end_longitude, seat_count = NULL)
}
\arguments{
\item{start_latitude}{Initial latitude.}

\item{start_longitude}{Initial longitude.}

\item{end_latitude}{Final latitude.}

\item{end_longitude}{Final longitude.}

\item{seat_count}{Number of passengers.}
}
\description{
Returns an estimated price range for each product offered at a given location
}
\examples{
\dontrun{
uber_estimate_price(start_latitude = 37.761492, start_longitude = -122.423941,
                    end_latitude = 37.775393, end_longitude = -122.417546)
}
}
\references{
\url{https://developer.uber.com/docs/rides/api/v1-estimates-price}
}

