% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_estimate_time}
\alias{uber_estimate_time}
\title{Time estimate}
\usage{
uber_estimate_time(start_latitude = NULL, start_longitude = NULL,
  start_address = NULL, product_id = NULL)
}
\arguments{
\item{start_latitude}{Initial latitude.}

\item{start_longitude}{Initial longitude.}

\item{start_address}{Initial address.}

\item{product_id}{Unique identifier representing a specific product.}
}
\description{
Returns ETAs for all products currently available at a given location. The ETA for each product is expressed in seconds.
}
\examples{
\dontrun{
uber_estimate_time(start_latitude = 37.761492, start_longitude = -122.423941)
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-estimates-time-get}
}

