% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_history}
\alias{uber_history}
\title{History}
\usage{
uber_history(limit = 5, offset = 0)
}
\arguments{
\item{limit}{Number of items to retrieve.}

\item{offset}{Offset the returned results.}
}
\description{
Returns data about a user's activity on Uber.
}
\details{
Requires an OAuth 2.0 token with the history or history_lite scope.
}
\examples{
\dontrun{
uber_history()
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v12-history-get}
}

