% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_me}
\alias{uber_me}
\title{User information}
\usage{
uber_me()
}
\description{
Returns information about the Uber user.
}
\details{
Requires an OAuth 2.0 token with the profile scope.
}
\examples{
\dontrun{
uber_me()
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-me-get}
}

