% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_payment_methods}
\alias{uber_payment_methods}
\title{Payment methods}
\usage{
uber_payment_methods()
}
\description{
Retrieve a list of the user's available payment methods.
}
\details{
Requires an OAuth 2.0 token with the request scope.
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-payment-methods-get}
}

