% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_places_get}
\alias{uber_places_get}
\title{Get place address}
\usage{
uber_places_get(place_id = c("home", "work"))
}
\arguments{
\item{place_id}{Either "home" or "work".}
}
\description{
Retrieve home and work addresses from an Uber user's profile.
}
\details{
Requires an OAuth 2.0 token with the places scope.
}
\examples{
\dontrun{
uber_places_get()
uber_places_get("home")
uber_places_get("work")
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-places-place_id-get}
}

