% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_places_put}
\alias{uber_places_put}
\title{Set place address}
\usage{
uber_places_put(place_id = c("home", "work"), address)
}
\arguments{
\item{place_id}{Either "home" or "work".}

\item{address}{Address to be assigned.}
}
\description{
Update home and work addresses for an Uber user's profile.
}
\details{
Requires an OAuth 2.0 token with the places scope.
}
\examples{
\dontrun{
uber_places_put("home", "115 St Andrews Dr, Durban North, 4051, South Africa")
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-places-place_id-put}
}

