% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_products}
\alias{uber_products}
\title{Products available}
\usage{
uber_products(latitude = NA, longitude = NA, product_id = NA)
}
\arguments{
\item{latitude}{Latitude of location.}

\item{longitude}{Longitude of location.}

\item{product_id}{Unique identifier representing a specific product.}
}
\description{
Returns information about the Uber products offered at a given location.
}
\examples{
\dontrun{
uber_products(latitude = -33.925278, longitude = 18.423889)
uber_products(product_id = "91901472-f30d-4614-8ba7-9fcc937cebf5")
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-products-get}
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-products-product_id-get}
}

