% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_requests}
\alias{uber_requests}
\title{Request a ride}
\usage{
uber_requests(start_latitude = NULL, start_longitude = NULL,
  end_latitude = NULL, end_longitude = NULL, start_address = NULL,
  end_address = NULL)
}
\arguments{
\item{start_latitude}{Initial latitude.}

\item{start_longitude}{Initial longitude.}

\item{end_latitude}{Final latitude.}

\item{end_longitude}{Final longitude.}

\item{start_address}{Initial address.}

\item{end_address}{Final address.}
}
\description{
Request a ride on the behalf of the authenticated user.
}
\details{
Requires an OAuth 2.0 token with the request scope.
}
\examples{
\dontrun{
uber_requests(start_address = "37 Beach Road, Mouille Point, Cape Town",
              end_address = "100 St Georges Mall, Cape Town City Centre, Cape Town")
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-requests-post}
}

