% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_requests_current}
\alias{uber_requests_current}
\title{Current request}
\usage{
uber_requests_current()
}
\description{
Retrieve details of the currently active request.
}
\details{
Requires an OAuth 2.0 token with the all_trips or request scope.
}
\examples{
\dontrun{
uber_requests_current()
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-requests-current-get}
}

