% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_requests_current_delete}
\alias{uber_requests_current_delete}
\title{Delete current request}
\usage{
uber_requests_current_delete()
}
\description{
Delete the currently active request.
}
\details{
Requires an OAuth 2.0 token with the request scope.
}
\examples{
\dontrun{
uber_requests_current_delete()
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-requests-current-delete}
}

