% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{uber_estimate_price}
\alias{uber_estimate_price}
\title{Price estimate}
\usage{
uber_estimate_price(start_latitude = NULL, start_longitude = NULL,
  end_latitude = NULL, end_longitude = NULL, start_address = NULL,
  end_address = NULL, seat_count = NULL)
}
\arguments{
\item{start_latitude}{Initial latitude.}

\item{start_longitude}{Initial longitude.}

\item{end_latitude}{Final latitude.}

\item{end_longitude}{Final longitude.}

\item{start_address}{Initial address.}

\item{end_address}{Final address.}

\item{seat_count}{Number of passengers.}
}
\description{
Returns an estimated price range for each product offered at a given location
}
\examples{
\dontrun{
uber_estimate_price(start_latitude = 37.761492, start_longitude = -122.423941,
                    end_latitude = 37.775393, end_longitude = -122.417546)
}
}
\references{
\url{https://developer.uber.com/docs/ride-requests/references/api/v1-estimates-price-get}
}

